Ver 0.10

+++ www/eliminar.php
@@ -12,13 +12,20 @@ $usuario = unserialize($_SESSION['usuario']);
 $rl = new RepositorioLibro();
 $libro = $rl->get_one($_GET['l']);
 
-if ($rl->delete($libro)) {
-$mensaje = "Libro eliminado con éxito";
-} else {
-$mensaje = "Error al eliminar la Libro";
-}
-header("Location: home.php?mensaje=$mensaje");
+if ($libro->getIdUsuario() != $usuario->getId()){
+    header('Error: La cuenta no pertenece al USUARIO');
+    }
 
+if($libro->getStock() != 0) {
+    header('Location: home.php?mensaje= El Stock disponible no puede ser mayor a 0');
+    } else {
+        if ($rl->delete($libro)) {
+            $mensaje = "Libro eliminado con éxito";
+            }else {
+            $mensaje = "Error al eliminar la Libro";
+            }
+            header("Location: home.php?mensaje=$mensaje");    
+    } 
 } else {
 header('Location: index.php');
 }
